pyproject.toml
[tool.poetry]
name = "mat-ue"
version = "0.1.0-alpha.1"
description = "Multi-agent trust estimation testing from unreal engine simulator"
authors = ["Spencer Hallyburton <spencer.hallyburton@duke.edu>"]
readme = "README.md"
license = "Apache-2.0"
homepage = "https://avstack.org"
repository = "https://github.com/cpsl-research/multi-agent-trust-ue"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
]
keywords = ["carla", "autonomy", "autonomous vehicles"]
include = ["CHANGELOG.md"]

[tool.poetry.dependencies]
python = "~3.10"
pygame = "^2.1"
cumm-cu117 = "<0.5.0"
spconv-cu117 = "~2.3.6"
carla = { url = "https://g-b0ef78.1d0d8d.03c0.data.globus.org/packages/carla/carla-0.9.13-cp310-cp310-linux_x86_64.whl" }
lib-avstack-carla = { path = "submodules/lib-avstack-carla", develop = true }
lib-avstack-api = { path = "submodules/lib-avstack-api", develop = true }
lib-avstack-core = { path = "submodules/lib-avstack-core", extras = ["percep"], develop = true }
mate = {path = "submodules/multi-agent-trust-estimator", develop = true }

[tool.poetry.group.test.dependencies]
snakeviz="^2.1.1"
pytest = "^7"
pytest-cov = "^4"
bandit = "^1.7"
black = {extras = ["jupyter"], version = "~22.12.0"}
mypy = "^0"
isort = "^5.11"
flake8 = "^5.0"
autoflake = "^1"

[build-system]
requires = [
    "poetry-core",
    "torch"
]
build-backend = "poetry.core.masonry.api"
